Removing react import path from WillChangeMotionValue
