+++ packages/framer-motion/src/value/use-will-change/__tests__/is.test.ts
@@ -1,5 +1,5 @@
 import { isWillChangeMotionValue } from "../is"
-import { WillChangeMotionValue } from "../index"
+import { WillChangeMotionValue } from "../WillChangeMotionValue"
 import { MotionValue } from "../.."
 
 describe("isWillChangeMotionValue", () => {
